Last commit (commit.final.final.final)
I've checked the project. I've commented the code relative to java swing, in order not to leave incomplete code.
I added a check for the price integer right at the insertion.
It is difficult to fix this problem, it would require me to change the entire structure, removing throw and leaving if. But this would go against the security requirements.

Boss, I'm tired (have a nice weekend)